% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayConverter_EPICtoMM.R
\name{arrayConverter_EPICtoMM}
\alias{arrayConverter_EPICtoMM}
\title{arrayConverter_EPICtoMM}
\usage{
arrayConverter_EPICtoMM(betaEPIC, MammalianList)
}
\arguments{
\item{betaEPIC}{A beta value dataframe (EPIC or 450k array) with column names as sample ID 
and row names as probe name.}

\item{MammalianList}{Vector contains probe names from Illumina HorvathMammalianMethylChip40 
BeadChip that need to be imputed.}
}
\value{
Imputed beta value matrix.
}
\description{
Impute mammalian methylation array data from human Infinium 
array data.The coefficients were generated based on a study of n=141 human 
blood samples that were profiled using both the mammalian array and the 
Illumina 450k array.  each mammalian CpG was fitted by penalized regressions.
}
\examples{
data('subGSE174422')
n_betas <- arrayConverter_EPICtoMM(betas, c('cg00249943', 'cg00250826', 'cg00292639'))

}
\author{
This is adopted from Lu, Ake T., et al.(PMID: 37563227)
}
