% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableClock.R
\name{availableClock}
\alias{availableClock}
\title{List all available clocks}
\usage{
availableClock()
}
\value{
A vector includes all available clock names
}
\description{
List all available clocks
}
\examples{
availableClock()

}
