% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiTOC2.R
\name{epiTOC2}
\alias{epiTOC2}
\title{epiTOC2}
\usage{
epiTOC2(betas, coefs, full_model = TRUE)
}
\arguments{
\item{betas}{A dataframe with column names as sample ID and row names as probe name,
betas = M / (M + U + 100).}

\item{coefs}{A dataframe contains coefficents for each probes, colnames as: delta, beta0.}

\item{full_model}{Default: TRUE, whether to use the full model; In the full model, we use the 
estimated methylation ground state values, whereas in the simplified model 
we assume that these are all zero. Reason for using the simplified model is 
because it could happen that some methylation beta-values in the data matrix 
are lower than the ground-state values, which in principle is not allowed. 
If this is the case, then more reliable estimates are provided by using the 
simplified model.}
}
\value{
A vector.
}
\description{
Estimate the cumulative number of stem-cell divisions in a sample using the 
epiTOC2 model.
}
\examples{
data(list='epiTOC2', package='dnaMethyAge')
data('subGSE174422') ## load example dataset, include: betas, info.
print(dim(betas))
# [1] 485577 8

res <- epiTOC2(betas, coefs, full_model=TRUE)
res2 <- epiTOC2(betas, coefs, full_model=FALSE)


}
\author{
This function is adopted from Andrew E Teschendorff, 2020 
(https://doi.org/10.1186/s13073-020-00752-3) by Yucheng Wang 
(wangyucheng511[at]gmail.com)  with modifications.
}
