% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanImputation.R
\name{meanImputation}
\alias{meanImputation}
\title{Imputate array samples, replace NA with sample mean or golden reference}
\usage{
meanImputation(mt, ref, cut_off = 0.9, only_ref_rows = TRUE)
}
\arguments{
\item{mt}{A numeric matrix with column names as sample ID and row names as probe name.}

\item{ref}{Golden reference, a vector consists of numeric values, in which every item 
should has a name, e.g. setNames(c(1,2,3), c('a', 'b', 'c'))}

\item{cut_off}{Default 0.9, if the ratio of samples with NA in a row greater than this cutoff,
then replace all samples with an identical reference value at that row.}

\item{only_ref_rows}{Default TRUE. TRUE: the final dataframe will only keep rows with names exist 
the in reference vector, FALSE: all rows with names presented in original 
dataframe and reference vector will be saved.}
}
\value{
A dataframe with all NA replaced.
}
\description{
Imputate array samples, replace NA with sample mean or golden reference
}
\examples{

mt <- data.frame(x1=c(1, NA, 3), x2=c(NA, 10, 15), x3=c(3, 5, NA), 
                  row.names=c('a', 'b', 'c'))

ref <- ref <- setNames(c(1, 2, 3, 4, 5), c('a', 'b', 'c', 'd', 'e'))

n_mt <- meanImputation(mt, ref)


}
