% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAccel.R
\name{getAccel}
\alias{getAccel}
\title{Calculate age acceleration.}
\usage{
getAccel(
  c_age,
  m_age,
  method = "Linear",
  do_plot = TRUE,
  title = "",
  point_color = NA,
  point_shape = NA,
  plot_accel = TRUE,
  simple = FALSE,
  plot_legend = TRUE,
  x_lim = c(0, 100),
  y_lim = c(0, 100),
  x_lab = "Chronological Age",
  y_lab = "mAge",
  y2_lab = "Age Acceleration",
  marker_cex = 0.7,
  axis_cex = 0.9,
  text_cex = 0.7,
  lab_cex = 1
)
}
\arguments{
\item{c_age}{A vector includes the chronological age of each sample.}

\item{m_age}{A vector includes the epigenetic ages with sample order the same with 'c_age'}

\item{method}{Default: 'Linear', methods to calculate age acceleration, available choices
include "None", "Linear", and "Loess".}

\item{do_plot}{Default: TRUE, whether to visualise the age acceleration results.}

\item{title}{Default: NA, set the title for the plot.}

\item{point_color}{Default: NA, define the color for the points in the plot, can be a character or
a vector of characters.}

\item{point_shape}{Default: NA, define the shape for the points in the plot, can be a single 
integer or a vector of characters.}

\item{plot_accel}{Default: TRUE, whether to plot the accel distribution.}

\item{plot_legend}{Default: TRUE, whether to plot the legend.}
}
\value{
A vector represents age acceleration in each sample.
}
\description{
Calculate age acceleration.
}
\examples{
c_age <- sample(1:100, 20)
m_age <- c_age + sample(1:10, 20, replace = TRUE)

## 1. calculate age acceleration
res <- getAccel(c_age, m_age, method='Linear', do_plot=FALSE)
print(res)

## 2. calculate age acceleration and visualise age accel results.
pdf('savename.pdf', width=4.3, height=6)
res <- getAccel(c_age, m_age, method='Linear', title='TEST')
dev.off()

}
