% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAccel.R
\name{getAccel}
\alias{getAccel}
\title{Calculate age acceleration.}
\usage{
getAccel(c_age, m_age, method = "Linear")
}
\arguments{
\item{c_age}{A vector includes the chronological age of each sample.}

\item{m_age}{A vector includes the epigenetic ages with sample order the same with 'c_age'}

\item{method}{Default: 'Linear', methods to calculate age acceleration, available choices
include "None", "Linear", and "Loess".}
}
\value{
A vector represents age acceleration in each sample.
}
\description{
Calculate age acceleration.
}
\examples{
c_age <- sample(1:100, 20)
m_age <- c_age + sample(1:10, 20, replace = TRUE)
res <- getAccel(c_age, m_age, method='Linear')
print(res)

}
