% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethyAge.R
\name{methyAge}
\alias{methyAge}
\title{Predict epigenetic age from DNA methylation data}
\usage{
methyAge(
  betas,
  clock = "HorvathS2013",
  age_info = NA,
  fit_method = "Linear",
  do_plot = TRUE,
  inputation = TRUE,
  simple_mode = FALSE,
  species = "Homo sapiens",
  MM_array = FALSE,
  use_cores = detectCores()
)
}
\arguments{
\item{betas}{A dataframe with column names as sample ID and row names as probe name,
beta = M / (M + U + 100).}

\item{clock}{Default: 'HorvathS2013', define the clock name to use, currently supported
clocks are 'HannumG2013', 'HorvathS2013', 'LevineM2018', and 'ZhangQ2019'.}

\item{age_info}{Default: NA, in order to calculate the age acceleration, you need to provide 'age_info'
with a dataframe which contains sample ID and age information.}

\item{fit_method}{Default: 'Linear', select a method to calculate age acceleration, avaliabe
choices include "None", "Linear", and "Loess".}

\item{do_plot}{Default: TRUE, whether to visualise the age acceleration results. Only valid 
when age_info is supplied with expected values.}

\item{inputation}{Default: TRUE, whether to make imputation to replace NA with sample mean or 
fixed reference, please refer to help(meanImputation) to know how imputations
are carried out in this function.}

\item{simple_mode}{Default: FALSE, whether not to perform data normalisation for the clock of
HorvathS2013.}

\item{species}{default: 'Homo sapiens', Latin name of the species studied, this is needed by 
clocks of 'LuA2023p2' and 'LuA2023p3'}

\item{MM_array}{Default: FALSE, is the input beta value matrix from Illumina 
HorvathMammalianMethylChip40?}

\item{use_cores}{An integer(e.g. 8) defines the number of cores to use when paralleling preprocess the
normalisation step for HorvathS2013. Only valid under Linux OS, default value 
is the maximum number of cores available.}
}
\value{
A dataframe includes predicated epigenetic ages, age acceleration is only 
returned when age_info is supplied with a dataframe contain sample age 
information.
}
\description{
Predict epigenetic age from DNA methylation data
}
\examples{

data('subGSE174422') ## load example dataset, include: betas, info.
print(dim(betas))
# [1] 485577 8

#### 1. Predict epigenetic age from DNA methylation data
horvath_age <- methyAge(betas, clock='HorvathS2013', simple_mode=TRUE) ## fast mode

## Reliable mode, beta values are subjected to fixed-reference based BMIQ 
## normalisation, same as Horvath's paper
horvath_age <- methyAge(betas, clock='HorvathS2013') 

hannum_age <- methyAge(betas, clock='HannumG2013')

pheno_age <- methyAge(betas, clock='LevineM2018')

zhang_age <- methyAge(betas, clock='ZhangQ2019')

Pan-mammalian_clock2 <- methyAge(betas, clock='LuA2023p2')

#### 2. Predict epigenetic age and calculate age acceleration
hannum_age <- methyAge(betas, clock='HannumG2013', age_info=info)
horvath_age <- methyAge(betas, clock='HorvathS2013', age_info=info, fit_method='Linear')
pheno_age <- methyAge(betas, clock='LevineM2018', age_info=info)
zhang_age <- methyAge(betas, clock='ZhangQ2019', age_info=info)

}
